Try to get the svn.exe (if installed) to read the current local svn revision of a file.
TortoiseProc from TortoiseSVN 1.9.x doesn't work with /endrev:-1 so try to get the local revision and pass that as a parameter.
If svn.exe is not installed then neither the /startrev not the /endrev parameters will be passed to display the dialog.

Also display the file version of TortoiseProc in the about dialog.

git-svn-id: svn://svn.code.sf.net/p/delphitsvnaddin/code/trunk@86 4dc0f469-f4c8-4329-b7b7-9aa88fb956bb